\name{BIC}
\alias{BIC.randomLCA}
\title{BIC for randomLCA object}
\description{
Returns BIC for a randomLCA object.
}
\usage{
\method{BIC}{randomLCA}(object, \ldots)
}
\arguments{
  \item{object}{randomLCA object}
  \item{\dots}{ additional argument; currently none is used. }
}
\value{
BIC
}
\author{Ken Beath}
\keyword{methods}