\name{class.probs}
\alias{class.probs}
\title{
Determines class probabilities for fitted model
}
\description{
The class probabililities for the model are returned.
}
\usage{
class.probs(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{randomLCA object}
}
\details{
Simply extracts the corresponding variable from the randomLCA object.
}
\value{
A vector of class probabilities for each class.
}
\author{Ken Beath}
\keyword{methods}
