\name{fitted}
\alias{fitted.randomLCA}
\title{fitteded values}
\description{
Extract fitteded values for randomLCA object
}
\usage{
\method{fitted}{randomLCA}(object, \ldots)
}
\arguments{
  \item{object}{randomLCA object}
  \item{\dots}{ additional argument; currently none is used. }
}
\value{
Vector of fitteded values corresponding to the observed frequencies.
}
\author{Ken Beath \email{ken@kjbeath.com.au}
}
\examples{
}
\keyword{methods}