\name{print.randomLCA}
\alias{print.randomLCA}
\title{print for randomLCA object}
\description{
Prints a randomLCA object. Prints summary.
}
\usage{
\method{print}{randomLCA}(x, \ldots)
}
\arguments{
  \item{x}{randomLCA object}
  \item{\dots}{ additional argument; currently none is used. }
}
\author{Ken Beath}
\keyword{methods}