\name{refit}
\alias{refit}
\alias{refit.randomLCA}
\title{Refit an randomLCA object}
\description{
Refits an randomLCA object using new data.
}
\usage{
\method{refit}{randomLCA}(object, newpatterns, newfreq, \ldots)
}
\arguments{
  \item{object}{randomLCA object}
  \item{newpatterns}{the new patterns that are to be fitted using the existing model}
  \item{newfreq}{the frequencies corresponding to the patterns if required}
  \item{\dots}{ additional argument; currently none is used. }
}
\details{
Not complete.
}
\value{
The fitted model to the new data.
}
\author{Ken Beath}
\keyword{methods}