\name{summary}
\alias{summary.randomLCA}
\title{Summary for randomLCA object}
\description{
Summarises the fit of a randomLCA object.
}
\usage{
\method{summary}{randomLCA}(object, \ldots)
}
\arguments{
  \item{object}{randomLCA object}
  \item{\dots}{ additional argument; currently none is used. }
}
\details{
Not complete.
}
\value{
  \item{log.Lik}{Log Likelihood}
  \item{AIC}{AIC}
  \item{BIC}{BIC}
}
\author{Ken Beath}
\examples{
}
\keyword{methods}