\name{AIC}
\alias{AIC.randomLCA}
\title{AIC for randomLCA object}
\description{
Returns AIC for a randomLCA object.
}
\usage{
\method{AIC}{randomLCA}(object, \ldots)
}
\arguments{
  \item{object}{randomLCA object}
  \item{\dots}{ additional argument; currently none is used. }
}
\details{
Not complete.
}
\value{
}
\author{Ken Beath}
\examples{
}
\keyword{methods}