\name{logLik}
\alias{logLik.randomLCA}
\title{log Likelikelihood for randomLCA object}
\description{
Returns log Likelihood for a randomLCA object.
}
\usage{
\method{logLik}{randomLCA}(object, \ldots)
}
\arguments{
  \item{object}{randomLCA object}
  \item{\dots}{ additional argument; currently none is used. }
}
\details{
Not complete.
}
\value{
}
\author{Ken Beath}
\examples{
}
\keyword{methods}