\name{simulate}
\alias{simulate.randomLCA}
\title{Simulate}
\description{
Simulate data from a fitted randomLCA model}
\usage{
\method{simulate}{randomLCA}(object, nsim, seed, \ldots)
}
\arguments{
  \item{object}{randomLCA object}
  \item{nsim}{number of data sets to be simulated}
  \item{seed}{random seed}
  \item{\ldots}{additional optional arguments.}
}
\details{
}
\value{
}
\author{Ken Beath}
\examples{
}
\keyword{methods}
